# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.LoginData:
    properties:
      email:
        type: string
      password:
        type: string
      tel_no:
        type: string
    type: object
...
  handlers.RegisterData:
    properties:
      hospital:
        $ref: '#/definitions/models.Hospital'
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.RegisterDataBody:
    properties:
      hospital:
        $ref: '#/definitions/models.HospitalBody'
      user:
        $ref: '#/definitions/models.UserBody'
    type: object
...
  handlers.UpdatePersonelBody:
    properties:
      job_type:
        type: string
      name:
        type: string
      polyclinic_id:
        type: integer
      surname:
        type: string
      tc_no:
        type: string
      tel_no:
        type: string
      title:
        type: string
      working_days:
        type: string
    type: object
  models.Hospital:
    properties:
      adressDetail:
        type: string
      city:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      district:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      polyclinics:
        items:
          $ref: '#/definitions/models.Polyclinic'
        type: array
      taxID:
        type: string
      telNo:
        type: string
      updatedAt:
        type: string
      user:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.HospitalBody:
    properties:
      adressDetail:
        type: string
      city:
        type: string
      district:
        type: string
      email:
        type: string
      name:
        type: string
      taxID:
        type: string
      telNo:
        type: string
    type: object
  models.Personel:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      hospital_id:
        type: integer
      id:
        type: integer
      job_type:
        type: string
      name:
        type: string
      polyclinic:
        $ref: '#/definitions/models.Polyclinic'
      polyclinic_id:
        type: integer
      surname:
        type: string
      tc_no:
        type: string
      tel_no:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      working_days:
        type: string
    type: object
  models.PersonelBody:
    properties:
      job_type:
        type: string
      name:
        type: string
      surname:
        type: string
      tc_no:
        type: string
      tel_no:
        type: string
      title:
        type: string
      working_days:
        type: string
    type: object
  models.Polyclinic:
    properties:
      city:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      district:
        type: string
      hospital:
        $ref: '#/definitions/models.Hospital'
      hospitalID:
        type: integer
      id:
        type: integer
      personels:
        items:
          $ref: '#/definitions/models.Personel'
        type: array
      polyclinic_name:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      hospital:
        $ref: '#/definitions/models.Hospital'
      hospitalID:
        type: integer
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        type: string
      surname:
        type: string
      tcNo:
        type: string
      telNo:
        type: string
      tokenVersion:
        type: integer
      updatedAt:
        type: string
    type: object
  models.UserBody:
    properties:
      email:
        type: string
      hospitalID:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        type: string
      surname:
        type: string
      tcNo:
        type: string
      telNo:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API for managing authentication
  title: Auth API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login a user and return a JWT token.You can login with email or
        tel_no
      parameters:
      - description: Login Data
        in: body
        name: loginData
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Login a user
      tags:
      - auth
  /personel:
    post:
      consumes:
      - application/json
      description: Add a new personel to the hospital
      parameters:
      - description: personel
        in: body
        name: personel
        required: true
        schema:
          $ref: '#/definitions/models.PersonelBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Personel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a new personel
      tags:
      - personel
  /personel/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing personel by ID
      parameters:
      - description: Personel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Personel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a personel
      tags:
      - personel
    get:
      consumes:
      - application/json
      description: Retrieve a single personel by ID
      parameters:
      - description: Personel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Personel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get personel by ID
      tags:
      - personel
    put:
      consumes:
      - application/json
      description: Update an existing personel's details.gorm.Model shouldnt be included
        here.(createdAt,updatedAt,deletedAt)
      parameters:
      - description: personel ID
        in: path
        name: id
        required: true
        type: integer
      - description: personel
        in: body
        name: personel
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePersonelBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Personel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a personel
      tags:
      - personel
...
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user and hospital,Ignore HospitalID in User
      parameters:
      - description: Register Data
        in: body
        name: registerData
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterDataBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RegisterData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"